Metadata pt. two
Added SemanticError metadata, though it still needs a great ammount of work. It saves a lot of time when debugging, tho, so it's good.
Messed arround with Issue 8, it doesn't manifest itself anymore, but it's still there.
I'll take some time rethinking String and its range representation, cause this bug may prove annoying in the future.
It's caused by String.range() returning a range of one character long Strings, which then return a range of Strings - themselves and so on.
Lastly, I got rid of the libedit untill I fix the unicode support. This commit fixes String.range() not respecting code points.