Macro and a new bug
Moved macro to a separate class for speed.
Figured out that the memory issues were in fact GDC issues, so these will probably be back in the Callable together with Closure.

Found a bug causing the interpreter to die with a stack overflow (tests/test3.asm), I'll have a look at this some time later this week.
It causes some symbols (like `assoc' from imports/core.asm) to enter an infinite call loop when a module is imported for a second time.